requirements.txt
streamlit
fastapi
uvicorn
uvicorn[standard]
python-dotenv
python-multipart
langchain
langchain_community
langchain_experimental
langchain_pinecone
langchain_ollama
langchain_core
langchain_openai
pinecone
docling
groq
streamlit-mic-recorder